refactor: move models to shared folder
