+++ src/app/search/search-results/search-results.component.ts
@@ -1,8 +1,8 @@
 import { CommonModule } from '@angular/common';
 import { Component } from '@angular/core';
 import { ChangeBorderColorDirective } from 'app/search/directives/change-border-color.directive';
-import { SearchItem } from 'app/search/models/search-item.model';
 import { SearchItemComponent } from 'app/search/search-item/search-item.component';
+import { SearchItem } from 'app/shared/models/search-item.model';
 import { VideoItemsService } from 'app/shared/services/video-items.service';
 
 @Component({
