+++ src/app/shared/models/search-item.model.ts
